refactor: Rename "concepts" to "elements"
