+++ common/src/main/java/com/wynntils/models/character/actionbar/PowderSpecialSegment.java
@@ -6,7 +6,7 @@ package com.wynntils.models.character.actionbar;
 
 import com.wynntils.handlers.actionbar.ActionBarPosition;
 import com.wynntils.handlers.actionbar.ActionBarSegment;
-import com.wynntils.models.concepts.Powder;
+import com.wynntils.models.elements.type.Powder;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
