+++ common/src/main/java/com/wynntils/models/elements/type/Element.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2023.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.concepts;
+package com.wynntils.models.elements.type;
 
 import com.wynntils.utils.StringUtils;
 import net.minecraft.ChatFormatting;
