+++ common/src/main/java/com/wynntils/models/elements/type/PowderProfile.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2022.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.concepts;
+package com.wynntils.models.elements.type;
 
 import java.util.Arrays;
 import java.util.Collection;
