+++ common/src/main/java/com/wynntils/models/gear/GearInfoRegistry.java
@@ -17,8 +17,8 @@ import com.wynntils.core.components.Models;
 import com.wynntils.core.net.Download;
 import com.wynntils.core.net.UrlId;
 import com.wynntils.models.character.type.ClassType;
-import com.wynntils.models.concepts.Element;
-import com.wynntils.models.concepts.Skill;
+import com.wynntils.models.elements.type.Element;
+import com.wynntils.models.elements.type.Skill;
 import com.wynntils.models.gear.type.GearAttackSpeed;
 import com.wynntils.models.gear.type.GearDropType;
 import com.wynntils.models.gear.type.GearInfo;
