+++ common/src/main/java/com/wynntils/models/gear/parsing/GearParser.java
@@ -8,8 +8,8 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import com.wynntils.core.WynntilsMod;
 import com.wynntils.core.components.Models;
-import com.wynntils.models.concepts.Powder;
-import com.wynntils.models.concepts.Skill;
+import com.wynntils.models.elements.type.Powder;
+import com.wynntils.models.elements.type.Skill;
 import com.wynntils.models.gear.type.GearInfo;
 import com.wynntils.models.gear.type.GearTier;
 import com.wynntils.models.gear.type.GearType;
