+++ common/src/main/java/com/wynntils/models/gear/tooltip/GearTooltipFooter.java
@@ -4,7 +4,7 @@
  */
 package com.wynntils.models.gear.tooltip;
 
-import com.wynntils.models.concepts.Powder;
+import com.wynntils.models.elements.type.Powder;
 import com.wynntils.models.gear.type.GearInfo;
 import com.wynntils.models.gear.type.GearInstance;
 import com.wynntils.models.gear.type.GearMajorId;
