+++ common/src/main/java/com/wynntils/models/gear/tooltip/GearTooltipHeader.java
@@ -6,8 +6,8 @@ package com.wynntils.models.gear.tooltip;
 
 import com.wynntils.core.components.Models;
 import com.wynntils.models.character.type.ClassType;
-import com.wynntils.models.concepts.Element;
-import com.wynntils.models.concepts.Skill;
+import com.wynntils.models.elements.type.Element;
+import com.wynntils.models.elements.type.Skill;
 import com.wynntils.models.gear.type.GearInfo;
 import com.wynntils.models.gear.type.GearInstance;
 import com.wynntils.models.gear.type.GearRequirements;
