+++ common/src/main/java/com/wynntils/models/gear/type/GearInstance.java
@@ -5,7 +5,7 @@
 package com.wynntils.models.gear.type;
 
 import com.wynntils.core.WynntilsMod;
-import com.wynntils.models.concepts.Powder;
+import com.wynntils.models.elements.type.Powder;
 import com.wynntils.models.stats.StatCalculator;
 import com.wynntils.models.stats.type.StatActualValue;
 import com.wynntils.models.stats.type.StatPossibleValues;
