+++ common/src/main/java/com/wynntils/models/gear/type/GearRequirements.java
@@ -5,7 +5,7 @@
 package com.wynntils.models.gear.type;
 
 import com.wynntils.models.character.type.ClassType;
-import com.wynntils.models.concepts.Skill;
+import com.wynntils.models.elements.type.Skill;
 import com.wynntils.utils.type.Pair;
 import java.util.List;
 import java.util.Optional;
