+++ common/src/main/java/com/wynntils/models/items/annotators/gui/SkillPointAnnotator.java
@@ -6,7 +6,7 @@ package com.wynntils.models.items.annotators.gui;
 
 import com.wynntils.handlers.item.ItemAnnotation;
 import com.wynntils.handlers.item.ItemAnnotator;
-import com.wynntils.models.concepts.Skill;
+import com.wynntils.models.elements.type.Skill;
 import com.wynntils.models.items.items.gui.SkillPointItem;
 import com.wynntils.utils.mc.LoreUtils;
 import java.util.regex.Matcher;
